+++ backend/src/main/java/com/example/metalTest/mantenimientoCorrectivo/domain/MantenimientoCorrectivo.java
@@ -33,7 +33,7 @@ public class MantenimientoCorrectivo {
     private Date fechaFin;
 
     @Column
-    private int tipofalla;
+    private short tipofalla;
 
     @Column
     private int horasProduccionAfectadas;
